requirements.txt
asgiref==3.3.4
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
Markdown==3.3.4
mysqlclient
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
django-heroku
django-cors-headers
html2text
python-dotenv
django-seed
django-rest-passwordreset
sib-api-v3-sdk
requests
moesifdjango
Pillow